Project/requirements.txt
streamlit
openai==0.28
qdrant-client[http]
tqdm
beautifulsoup4
python-dotenv